requirements.txt
Django==1.11.3
django-ckeditor==5.3.0
django-js-asset==0.1.1
django-mptt==0.8.7
django-suit==0.2.25
olefile==0.44
Pillow==4.2.1
pytz==2017.2